Add postcss and stylelint
Since we use webpack with multi entrypoints and want's to have the styling very decoupled from each component we need to have the whole pibeline kinda decoupled. This is to avoid duplication of code that we ship.
Two applications might end up having the same atoms css embedded into their app which we do not want.

We still want to use scss for familiarity but we can't just go with node-sass and call it a day. We also want autoprefixer etc. We could do this via gulp but that would add a new dependency on top of node-sass and postcss. We do it with vanilla postcss-cli now.

Stylelint is being pulled in to maintain simple name spacing rules and make the code consistent. A lot of different people will get their hands on this. Might try to gate it a bit up front.

All of these tools have some ideas of how you are supposed to be using them and since we have quite a special use case we more or less hack it together at the end with the use of concurrently and chokidar.

yarn dev is still enough to get up and developing and yarn build should still produce delpoyable artifacts.