Set token for development.
When developing we want to be able to set a token that can be used for testing.
These tokens lasts for 30 days so manual switching should be okay.
A warning will appear if there is no .token file and/or if the .token file is empty.

We load in the .token file in on the webpack bootstrap level and when Storybook is starting up.
From then on we can access it by reffing process.env["ddb-token"].
We only want to ref it once in our setToken function that is called on the "client" of Storybook.
This takes the now string transformed process.env["ddb-token"] and injects it into localStorage.