Restructure mounting.
The only thing the new addMount does is to add the application and it's name as a key value pair to the global scope object named as ddbApps.

This is later down the line used in corresponcenze with the search of containers to figure out which apps to mount and where to mount them.