Implement ddbReact object and auth check.
Here we migrate the old flats structure into a container object that in context makes a bit more sense.
The apps are now located in window.ddbReact.apps etc.

Besides that we introduce userAuthenticated which should be set by the server. We use this for adjustments to look and navigational adjustments.