Add support for injecting service urls through props
This will allow users of our components to determine which service
they want to integrate with e.g. for switching between staging and
production versions of the service.

By using props we are using the same approach as we use for other
customizations. To retain current behavior the urls default to 
non-production versions.

For this to work we have to move service instance creation inside our
hooks. This is a bit less elegant but if we only move in inside our
Component function React will keep rerendering the components.

Add a knob for controlling service urls in the Storybook.