Fix list content sizing in IE11
In the current form list items with long titels have a very long
which may overlap with the buttons to the right.

This probably has to do with the fact that we use unset for max-width
which is not supported by IE11. If we set it to none all items will
be too wide.

This is then addressed by setting max-width to 100% instead of auto.