Remove empty objects from OpenPlatform work responses
OpenPlatform may return empty objects for materials which have no
corresponding metadata. They end up looking weird in the checklist.

In general we are not interested in processing such objects so remove
them from the response. It is an exercise for the caller to determine
which if any requested pids have not yielded corresponding meta data.