Correct named inline functions.
Since we have enabled the AirBnB prefer-arrow functions rule some of out nice to have named inline functions would get lost in unnamed inline arrow functions.
Therefore we take most of those inline named functions and make them seperate independent functions in the right contextual space.
This allows for an easier named debugging down the road where we won't have to track an error to the top function to figure out what is going wrong but can track it to the sub function.