Allow Codecov to fail
The Codecov uploader requires a token which is typically stored in a
GitHub secret within the repository. However these secrets are not 
used for PRs of forks. This causes upload and thus our entire build
to fail.

Codecov seems to be working on tokenless uploads but there is no good
way to solve this at the moment. Allow Codecov to fail while we wait.
At least this will allow code coverage generation when working within
the main repository.

https://github.com/codecov/codecov-action/issues/29#issuecomment-581540357