Add support for specifying multiple material ids when ordering
This is useful if several different versions of a work are considered
equivalent and the patron allows the library to figure out how the
order can fulfilled most effectively.

Update the application to accept an array of id strings or a single
string containing multiple comma-separated ids. The latter version is
used when using the naive application mount with props passed as data
attributes.

Transformation to ensure that we only work with arrays happen as 
early as possible while avoiding update loops due to changing hook
dependencies. Arguments, proptypes etc. are updated down the call 
stack to work to only work with arrays.

Update storybook to also use a single string containing multiple ids.
To provide a relevant example we have to replace Star Wars with
Harry Potter which exists in multiple editions. Sorry.