A set of materials can be ordered if a single material can be ordered
When passing an array of supposedly equivalent materials we consider
the set orderable if at least one of the materials can be ordered.

When the user places an order using these ids it will be up to the 
library system to determine how best to fulfill the order. Materials
which cannot be ordered should be disregarded in this process. 

Update tests to take multiple availability responses into account.