Avoid prop type error when a cover does not have a source
The url-prop-type library does not recognize the empty string as a
valid url. Consequently we change it to null. Alternatively we could
also have set it to a string it allows e.g. /. That would not yield a
valid image though.

This occurs when the cover is in its initial state.