Optionally create version.json file when building
This allows us to determine exactly what version is running where in
an environment.

Use the VersionFile webpack plugin to render a version.json file from
a template.

The way this project is included in other systems we cannot rely on
information provided by package.json so instead we use environment 
variables.