Exclude materials with too small covers from Related Materials app
We require a certain size of the cover images we retrieve. If the 
image is not large enough it may not display right on certain
devices. Consequently they should be removed for now.

Without this a related material will have a cover image with null as
src. This is just displayed as a blank space.

Expand tests accordingly.