Return related materials ordered by query instead of covers
In the current version we retrieve object ids from search result,
pass these to the cover service in the same order and merge material 
data intro the provided results.

However practice has shown that the cover service does not
necessarily preserve the order of the requested ids in the response.
This leads to a situation where the related materials will not have
the same order as in the linked search result. 

To preserve the order of the materials from the search result we 
merge cover data into the materials instead and filter away any 
materials without a cover.