Improve query generation for related materials
The current query does not sufficiently handle phrases and multiple
values. Consequently we have to update it for more advanced query
building.

We now use , as a separator between parts of a strong instead of 
passing the string directly to the query.

Update our storybook with more elaborate field titles and example
values that fit the updated pattern accordingly.