Do not render component if there are no related materials to show
If we do not locate any related materials then our component will
show a title and a link after loading. That looks somewhat silly.

We do not have a proper way to handle this situation by providing a
message to the end user so instead we just hide the component by
not rendering anything.

To do this we introduce the new “empty” status and set this if we
are done loading but still have no related materials.