Creat Availability Label component and use clsx library
Availability label is used to display whether a manifestation is
available for loaning. If it is, the label can be wrapped in a link.
This functionality will be used later in the project, where we redirect
users on click of the availability-label.

We introduce clsx npm package for conditional class rendering on
elements. We agreed that clsx is a more elegant solution to conditional
classes rendering than referencing functions from elements.