Readd functionality that sets library token as user token
..if library token is not available

We maintain the former functionality that sets the library token
to the user token if the library token is not available.
In this way it is both possible to use storybook by:
- Logging in
- Just set the library token.

Note that the manually set library token will take presedence
above the user token coming from the authentication.