Create Search header app
This app serves as an assembly of the search bar, search string
suggestion, material suggestion, and topic suggestion components.
It can be mounted into other projects, like for example dpl-cms.

It can redirect to different url's based on the variable that is given
to the app. This ensures it can operate when input into different
domains.