Update Graphql schema for the autosuggest component
Before we retreived id's on the type Work, but that is currently not
necessary. This also caused some issues when calling for data from the
API. The API service seems to be unfinished, and some of the data is
missing. Among this missing data are the id's for type Work. Due to this
whenever we would otherwise receive a Work with a null id the API
returns an error message and no data.
The graphql.schema was updated based on these changes using the
codegen:dbc:gateway command.

"Suggestion dropdown" component to "autosuggest dropdown" for more
accurate naming and the graphql query follows in this trend.