Introduce the autosuggest-text subcomponent
This component is going to be set into the main autosuggest component as
one of the three parts that make it up. It shows a list of item that are
strings. As we use Downshift to tackle the autosuggest, this component
utilizes some of downshift props to manage overall autosuggest state.