Add Downshift to the seach bar component and its story
As a part of adding downshift to our search header app, we now provide
new (downshift) props to the search bar component. This needed to be
accounted for in the component's story as well so that it can run and
show the component even without the search header app as its parent.
In the current time we still need to address some eslint erros and add
TS types to downshift props.