Update search header app entry props and their naming to end with "Text"
We add three new entry props for autosuggest-text as they contain static
texts that need to be translated/changed by the libraries from site to
site.

The input placeholder app prop now also ends in "Text" to comply with
variable prop naming service that we are currently developing. This
service looks for all the app props that end with "Text".