Use determineSuggestionType function to avoid repeating code
In the search header app there was an instance where we wanted to find
out what type of a suggestion we are currently working with, because
there are three types. This was repeated in the code, so it made sense
to make a separate function out of it that sorts things out instead of
repeating the code with slight adjustments for different contexts.