Use material component in material autosuggest & update the query
Before we just used static HTML structure as a placeholder. Now that the
material component is implemented we use it instead - so that we get the
right work covers for each work in the autosuggest.
We also had to update our autosuggest graphql query to retreive pid on
each work instead of the works' regular ids because the cover service
data calls only accept pid format, not the regular id format.