Go through the autosuggest items in the right order when using arrows
The probelm here was that when we got works back as suggestions any
earlier than on position 7, 8, and 9 in the suggestions array, using the
arrow keys to navigate the dropdown was buggy.
We needed to fix data division between components and handling of the
highlighted item logic in order to solve it. Now The first item from the
top visually will be highlighted first regardless of whether any works
came in before it from the API data call.