Manually redirect to search result page from autosuggest on enter hit
Before, we just updated the search field on enter (selection of one of
the suggested items), which was not the desired behavior. Now we can
distinguish between whether we update the field by writing in it, or by
browsing through the suggestions with arrow keys, or whether the user
has selected an item using enter - in which case we simulate a classic
form submission protocol.