Move Downshift combobox props from form to input wrapper
Integrating the current version of the autocomplete shows an
accessibility problem: The combobox aria role is added to the form
element where it is not allowed.

According to the Downshift documentation the combobox props should
be added to div which wraps search bar and autocomplete. This is in
accordance with documentation and examples for getInputProps for 
useCombobox() (https://www.downshift-js.com/use-combobox).

To ensure that the design stays the same we can drop the outer
wrapper div for the form and just bump classes to the form and inner
div instead.