Use dynamic data for material description + introduce HorizontalTermLine
We used to make a separate data call to get data for material
description, but now we added it into the work call used in the material
app component. Thus we had to resctructure the component a little bit.
As a part of this restructuring we create HorizontalTermLine component
that is reused in material description component to avoid repetition.