Add material app entry text props: loginToSeeReview & cantViewReview
These both end in "Text" so that they're saved into the redux state and
can be reused accross the app without prop drilling. They are going to
be used in infomedia reviews, where the user needs to be logged in and
their account has to be connected to a specific municipality.