Adjust query typing and navigation waiting in search header test
Currently tests seems to fail due to race conditions if Downshift
items have not been rendered before the user/test begins to use
arrow keys to navigate between them.

To avoid this we move the waiting to be after entering the search 
query but before using arrow keys to navigate.