Wait for actual instead of fixed time when testing search header
The relationship between typing in the search field and Downshift
displaying suggested queries is tricky to get right.

Previously we forced a specific time window. This can be replaced
with waiting for the autocomplete results to contain actual data.

This speeds up testing and conforms with Cypress best practices about
not using wait().