Add support for code coverage when using TypeScript
The current setup was established when the project was using plain
JavaScript. After the migration to TypeScript only coverage for the
remaining JavaScript files are analyzed.

To add support for TypeScript I have followed guidelines at 
https://github.com/lluia/cypress-typescript-coverage-example

This means we have to add a few packages to make our code coverage
engine Instanbul and NYC understand TypeScript. 
https://github.com/istanbuljs/istanbuljs/tree/master/packages/nyc-config-typescript 

We also reset our configuration to something plain as recommended by
@istanbuljs/nyc-config-typescript.