Combine integration test and code coverage in single workflow
We want the same process to run for Cypress tests and code coverage
generation on both pull requests and pushes (so we have something to
compare against).

Instead of duplicating the logic between separate workflows we might
as well combine them into one.

This means that we will upload screenshots and videos for both pushes
and pull requests. That should not make any difference.

Rename workflow name accordingly.