Run integration tests on all pushes
There does not seem to be a reason to only run it on specific
branches.

This is not something that is suggested in Codecov content
https://github.com/codecov/codecov-action
https://about.codecov.io/blog/how-to-set-up-codecov-with-c-and-github-actions/

It means that our tests will run more often but that should not make
a great difference anyway compared to the current situation where we
have to ensure that the setup matches our development workflow 1-1
which has also resulted in Codecov not running for a while.