Restructure development setup documentation
Move token information to the top as these are steps you will need
to perform before you can install the project with the design system.

Restructure steps with and without Docker

Move information about the Adgangsplatformen story to the part of the
documentation regarding access tokens as that is where tokens and
this story is mentioned otherwise.