Break up test of loan modal contents into individual tests
The current implementation with all elements in a single line is
brittle as it depends on how browsers handles whitespace, how
elements are ordered etc.

The breakup should make test test more stable and readable.

Using cy.log to show the relevance of each assertion would be nice
but unfortunately it is not chainable so we will have to do with
comments.