Add continual eslinting during development
This should make it easier to catch linting problems during development.

Developers may have integration with eslint directly in their IDE but
practice has shown that some issues still make it to pull requests.

To support this we add a new script for watching/linting of JS with
Chokidar and another new script for combined watching/linting with
Concurrently.

This new script replaces the current watching/linting of SCSS in the
general dev script.

For this to work in Docker we have to add the SHELL environment. It is required
by Chokidar and is not available in the default Node image.