Add lodash, its types & babel-plugin-lodash npm packages to the project
Summary of the talk:
There is a general need for utility functions that work with objects, as
the current out of the box object value filtering possibilities are not
ideal, the produced code can be quite confusing and the need for them
spans around the whole project. As such, we decided to add the full
lodash package to the project. We previously agreed on only importing
specific parts of the library, but now found out that lodash themselves
don't recommend that
(link for reference: https://lodash.com/per-method-packages) and this
method of working with lodash would increase the bundle size of the
project. Instead, we limit the bundle size by using the
lodash-recommended package - babel-plugin-lodash that only imports
specific methods used in components. Please refer to the link above for
further explanation.