Change DropDown props + Change logic for formatValuesToDropdown
DropDown title is change to label and value now need to be a string. This is more descriptive naming, and since select value can only be a string it must not be unknown.

formatValuesToDropdown now also accepts the facet name because we need to use it to know if a term is on the filter list