Set develop as the default branch for Codecov code coverage reporting
We have decided to drop release/* branches for continual development and
move this to a permanent develop branch.

The admin UI is configured to use release/4 and is filled with many
other branches. For some reason develop does not show up. Lets hope
defining it in code will work. It is preferable to have it in code
anyways.