Add Logic for orderDigitalCopy
React query is used to POST xml request and handle the status of the response for order digital copy.
Be aware that to be able to use this function cors must be turned off,
Here is a guide to open a chrome browser without cors https://keepforyourself.com/tools/how-to-run-chrome-without-cors/

DigitalModalSuccess + DigitalModalError components have been introduced to handle success or error,
state for that will also be handled by react query.