Introduce the reusable Dropdown in places where we used select before
Makes the code more readable and easier to maintain.

A dropdown can only have a single selected option so this is better controlled through the single defaultValue prop for the dropdown instead of on each option.

Instead of having a selected value on the Option Type, we now use the defaultValue to check inside the Dropdown if defaultValue is equal to the value of an option.

The consequence of using Option Type for placeholder is that we now need to set a value. This will in most cases be an empty string