Add hasCorrectSource + Check if material is eReolen to show Availability
hasCorrectSource takes in two arguments: desiredSource, which is an array of strings, and manifestation, which is an object of type Manifestation. The function returns a boolean value indicating whether the source property of the manifestation object includes at least one of the strings in the desiredSource array. If the manifestation object does not have a source property, the function will return false.