Add numberInQueue to StockAndReservationInfo
numberInQueue is an optional prop that specifies the number of users currently in a queue waiting for materials. If the numberInQueue prop is truthy, the component will display a message indicating the number of users in the queue (e.g. "3 users in queue"). If the numberInQueue prop is falsy, the component will not display any message about the number of users in the queue.