Introduce the concept of a main "find-on-shelf" modal on material page
We need to be able to show a find-on-shelf modal for every work mani-
festation, but also one for all manifestations of a currently selected
material type. This is the "main" version of the modal.
The find on shelf button also needs to be able to distinguish and open
the modal.