Introduce the concept of a main reservation modal on material page
We need to be able to show a reservation modal for every work mani-
festation, but also one for all manifestations of a currently selected
material type. This is the "main" version of the modal.
The reservation button also needs to be able to distinguish and open
the modal.
We only create these special modals for physical material types, as
I assume that materials in digital form only have one edition.