Add getManifestationsOrderByTypeAndYear
'getManifestationsOrderByTypeAndYear' is a function that takes an array of Manifestation objects as an argument. It uses two helper functions, orderManifestationsByYear and groupBy, to sort the array of Manifestation objects by year and then by material type. It returns a new array with the Manifestation objects sorted alphabetically by material type and then by year.