Rename getReleaseYear -> getReleaseYearSearchResult & use helpers in it
This helper function is specific to the search result page where
selecting the release year of a work depends on whether it is a
fictional work or not. This is not how it works on the material page.
So I renamed the function + used other helpers that already exist in it.