Adjust material, material reviews & all subcomponents based on new API
Based on the new changes in the API the graphQL fragments and data
structure have been adjusted - now we needed to finish these adjustments
by adding the changes to the react components. One bigger change is that
we now call for review data only if the disclosure is opened, as opposed
to getting it as a part of the work data on pageload.