Use 'usePlaceCopyMutation' instead of 'useDplDasDigitalArticleOrderPOS'
Replaces the usage of 'useDplDasDigitalArticleOrderPOS' with 'usePlaceCopyMutation' to order a digital copy. 'usePlaceCopyMutation' returns a status, which we use 'getResponseMessage' to determine the translation string to show.

'getResponseMessage' takes two parameters, articleResponse and t. The function checks if articleResponse is defined, and if so, it creates an object with the keys being the different statuses of 'CopyRequestStatus' and the values being strings returned from the useTextFunction. Finally, it returns the value of the key in the object that matches the status of articleResponse.