Refactor `getFirstMaterialTypeFromFilters`
Refactored code to improve code reliability by updating the access of materialTypeFilter using the FacetField.MaterialTypes enum instead of directly accessing the materialTypes property. Added a truthy check for materialTypeFilter before checking if it's included in allMaterialTypes to prevent errors related to undefined values. Additionally, incorporated the head function from the Lodash library to retrieve the first element of the Object.keys array and ensure that materialTypeFilter is assigned a value to avoid any potential errors related to accessing an undefined array index.