Refactor `usePager` + Use `isLoading` from react-query
This change makes the ResultPager component more flexible by allowing it to be aware of its loading state. The isLoading prop is optional and allows the component that uses the usePager hook to control the loading state of the ResultPager. This improves the user experience by providing visual feedback to the user when the component is loading.