Simplify `Heading` level selection logic
Removing unnecessary conditional logic. Instead of using an object to map heading levels to HTML tags, we simply use the level prop as the HTML tag name, since the level prop is already restricted to valid heading levels through the HeadingLevelType type.

the `level` prop is changed to be required since it should always be set for a Heading