Add 'aria-labelledby' to `Link` + Change name 'description' to 'altText'
When the 'coverSrc' property in the `Cover` component is invalid, the image won't be displayed, leading to an empty link. To address this, we use the 'aria-labelledby' attribute to associate the empty link with the title of the search result item. We generate a unique identifier for the title using the useId() hook, which is assigned to the id attribute of the title and used as the aria-labelledby value for the `Link`, improving accessibility.

Additionally, the description prop in the Cover component is renamed to altText.