Don't open ReservationModal modal if the user is blocked.
When an unauthenticated user clicks on the Reserve button, the user is sent to the login platform. The user is then sent back with a URL containing `&modal=reservation-modal-********`, which opens the modal. This is not desired when a user is blocked, therefore I have added a check `isUserAllowed` before the modal is rendered.

In addition, I have added `userIsBlocked` and `userIsAllowed` to make the code more readable and maintainable.