Add `useGetHoldings`
Introduces a new custom hook, `useGetHoldings`, which serves as a wrapper around the existing `useGetHoldingsV3` hook. The purpose of this new hook is to handle cases where 'blacklistedPickupBranchesConfig' needs to be considered while fetching availability data. By utilizing the `getBlacklistedArgs` function, this wrapper ensures that blacklisted branches are appropriately excluded from the query results.