DDFSOEG-508 Check user is logged in before fetching patron information
If there is no user token (the user is not logged in), the FBS fetcher will send the library token instead. This will not work when trying to fetch 'https://fbs-openplatform.dbc.dk/external/agencyid/patrons/patronid/v2', and after 3 attempts, it will throw an error. Therefore, I have added a check for if there is a user before every patron information fetch.