Add Vitest in source testing
While creating tests for some link component functionality I ran into an
issue with testing/mocking a function that originally was an unexported function.
There are other questionable hacks you can do
that is not a part of the Vitest recommended practices.
However does Vitest have the in source testing concept that solves the issue.
I suggest we make use of that.