Update test for filmstriben buttons
Our previous work no longer contains a movie so we replace it with a
new one, Underverden.

Underverden is a somewhat older movie from 2017 so we expect rights to 
be available for some time.

The movie is available in multiple physical and digital formats so
we have to click the right one to ensure that the expected type is
shown.