Fix gh action cypress image to use chrome 116
Because of errors in the communication between Cypress and Chrome 117+
we need to make sure that we run Chrome 116.
This is a temporary solution and we should upgrade the cypress action
in the future as a part of the work involved in upgrading Cypress.

The GitHub issue that lead us to the problem:
https://github.com/cypress-io/cypress/issues/27804