Feedback mesages for group modal loan renewal
*Group modal renewal feedback message*
This commit adds feedback to a patron when trying to bulk renew loans in
a group modal.

*General Message component*
This commits also introduces a way handling messages in a general
manner. In the first version of this there are two key components:
Message and ModalMessage. The Message is quite barebone and just
introduces the plain setup with a title a subtitle and some optional
markup. The ModalMessage component extends the Message component and is
adding a cta button and focustrap.

*Test*
The `filterRenewResponseData` function is being tested

*Handle which renewals went well*
Since the renewal response gives us various responses according to which
renewals that went well and which that did not, we can reflect that in
our modal messages. Also if we get a 200 back from the renewal request
but no loans could be renewed, we need to inform the patron about that.