Separate advanced search object and search query handling
We want to separate between two modes of searching:

1. Combining queries, indexes, operators and filters through our
   advanced search form (search object)
2. Entering a direct CQL query (search query)

To sepatate between these two states we introduce two states for the
advanced search form to correspond with each of them.

To handle this current structures for rows and filters are combined into
a search query type and all code is updated accordingly.

Since translation from search query to CQL now needs to occur at two
levels - preview and final search rendering - we move the helpers for
managing this to a shared utility type which corresponds with our types.