Support setting initially selected (default) options in Multiselect
The list of these are defined as a separate prop. While it would have
been preferable to define a selected property on each item but it proved
to be tricky to keep the value of this property in sync with user
interactions with Downshift.

Update our story and tests accordingly.