Stackable material & reservation material's children should handle click
So the problem was that "stackable material" and "reservation material"
are cards that need to be clickable, but they also have elements inside
that need to be clickable, sometimes with different effects.
We now make sure that the children elements handle the click events and
don't trigger the parent card onClick event handler.