Link to ereol material page from reservation details for ereol materials
Eg. a user has a reservation for the Bible from the Ereol platform -
upon opening the reservation details modal and clicking on the "Go to
Ereolen" external link button, they will be redirected to the Bible work
page on the Ereol platform. In the case of this access data not being
available in the FBI API the user is redirected to the Ereol home page
instead.