Get manifestation data by best representation for parallel reservations
Use our new query to retrieve data from the best representation of the
owner work of one of the faust ids in our parallel reservation.

In general our parallel reservation data structure only contain a single
faust id (from the first reservation in the group). This may seem liked
a problem but based on the assumption that any faust id in the group 
should return the same owner work which should return the same best 
representation this should actually be OK.

Since parallel reservations may appear in different locations throughout
the app handling is added to the fetchMaterial() higher order component.

We can the utilize the fact that both hooks for retrieving material data
use the same fragment.

Expand the reservation list Cypress test to show that data fetching for
parallel reservations works.