Update materialGridManual and MaterialGrid
This updates the materialGrid to expect an array of
MaterialGridItemProps.

This has been discussed wether it should be a general type used for all the components using the recommendedMaterial. I am still unsure, but i thought it was cleaner to keep it like this.

Feel free to comment your thoughts on this.

MaterialGridManual is updated to now parse the data into material objects containing a workID and optional materialType.

DDFFORM-369