Add `selectMirror` and Start/End Time to `OpeningHoursEditorEventContent`
With these enhancements, users will now see the time displayed when selecting time slots by dragging in the calendar.

Additionally, I have implemented some temporary inline CSS styles. It is important to note that these are provisional and should be replaced with CSS classes from the design system in a future update.