Differenciate between blacklistBranchesInstantLoan & pickupBlacklist
In the reservation modal - when changing between pickup branches, we
didn't ask for the correct configuration - forgetting
blacklistedPickupBranchesConfig branches. Now we add those to the
instant blacklisted branches, and as a result when an administrator
changes this config in the backend, the changes take effect in the
reservation modal as well.