Add form elements for event editing in `Dialog`
This will present the relevant values for the current event in input fields where the user can edit them.

There is a small change to the placeholder function `handleEventEditing` that now alerts the complete event data that will later be sent to the API.

I have removed `handleEventClick` as it is now unnecessary.