Add `withConfig` to handle props from dpl-cms
We use useConfig to manage props in order to handle arrays and number props. This is not the optimal solution since useConfig is starting to get a bit messy as, for example, these props here have nothing to do with config. We must deal with this another time.

I have removed GuardedApp since it is only related to login

- Implement `useWireMockStartDate` Boolean
This boolean aligns the full calendar start date with our wiremock data.

Also renamed the mocked data