Update fee-list.test.ts cypress test string.
In a previous commit related to OpeningHours, a global locale is set to danish.

This change made this test fail as the expected string is hardcoded to english.

For now the string is manually changed.

In the future, we should change the tests to run in this locale specifically

DDFFORM-417