Delete CMS model before generating a new one with Orval
This allows us to remove any part of the model that is no longer
relevant - either because the specification (or portentially Orval)
has been updated.

Orval has a clean configuration option that we could use for this
https://orval.dev/reference/configuration/output#clean

Unfortunately it does not seem to work properly. It also deletes our
custom mutator code. It seems to be a known issue:
https://github.com/anymaniax/orval/issues/333

For now we just delete the code by hand.