Move internal q state handling to custom var and handler
This q internal state is used to store the query string
without requesting the search service. This is used to keep the query
string in the search field when the user clicks on the autosuggest
item.
We also want to make sure that we don't rerender the component when
the user clicks on the autosuggest item, because it confuses
screen readers. So that is why we are NOT using `setState`.