Create MaterialSearchInputs & useUpdateFields
MaterialSearchInputs is the sub component for the search input and select input used in the app.

useUpdateFields is a custom hook that will listen to changes on these inputs, and update both the hidden/drupal fields that is used to save the values, as well as updating and clearing the relevant states related to displaying relevant data based on these states.

DDFFORM-611