Create MaterialSearchPreview & useGetSelectedWork
MaterialSearchPreview is the sub component used for displaying the selected material based. This can either be already selected, or updated by selecting an item from the list.

useGetSelectedWork is a custom hook responsible for fetching the material based on the selected material and optionally selected materialType.
it is also responsible for returning availableMaterialTypes that are present based on the selected material.

DDFFORM-611