Create basic details for materials based on FBS information
This can the be used as fallback information for physical materials in case
these cannot be retrieved from FBI in the material fetch higher-order component.

Without this then the material will not be rendered at all.

One known situation where this occurs is interlibrary loans from the Royal
Danish Library.