Update / Refactor materialSearch story to reflect changes in the approach of retrieving the workId and type through the DOM.
Also extended the stories to have with and without previously selected values.

I have not included an a story where the DOM elements could not be found, even tho i have added an error. There is already a lot of complexity, but if you think that i should add this aswell then comment it.

DDFFORM-833