Conditionally render material type in selectable lists
When using basic details from FBS the material type will not be
provided at the moment.

As such we will not have anything to render. Instead of rendering a
blank box as we do now we omit the element entirely.

No workarounds are required to ensure that the layout stays
consistent.