Introduce naming exceptions in generated graphql code.
Since the FBI schema is not consquent in the naming of the various
objects it conflicts with out Pascal casing of our typescript objects
(which is standard).

Sadly we need to accommodate the various oddities in naming of the
objects in the graphql schema, by renaming some of the conflicting
names.